requirements.txt
Flask
Werkzeug
sqlite3
Pillow
pandas
reportlab